Add FileTypes filter
