+++ src/filter/mod.rs
@@ -3,6 +3,7 @@ pub use self::time::TimeFilter;
 
 pub use self::common::Filter;
 pub use self::extensions::Extensions;
+pub use self::filetypes::FileTypes;
 pub use self::min_depth::MinDepth;
 pub use self::regex_match::RegexMatch;
 
@@ -11,6 +12,7 @@ pub use self::owner::OwnerFilter;
 
 mod common;
 mod extensions;
+mod filetypes;
 mod min_depth;
 mod regex_match;
