+++ src/main.rs
@@ -3,7 +3,6 @@ mod error;
 mod exec;
 mod exit_codes;
 mod filesystem;
-mod filetypes;
 mod filter;
 mod options;
 mod output;
@@ -26,10 +25,9 @@ use regex::bytes::{RegexBuilder, RegexSetBuilder};
 use crate::error::print_error;
 use crate::exec::CommandTemplate;
 use crate::exit_codes::ExitCode;
-use crate::filetypes::FileTypes;
 #[cfg(unix)]
 use crate::filter::OwnerFilter;
-use crate::filter::{SizeFilter, TimeFilter};
+use crate::filter::{FileTypes, SizeFilter, TimeFilter};
 use crate::options::Options;
 use crate::regex_helper::{pattern_has_uppercase_char, pattern_matches_strings_with_leading_dot};
